+++ nengo/__init__.py
@@ -52,7 +52,7 @@ from .probe import Probe
 from .rc import rc, RC_DEFAULTS
 from .simulator import Simulator
 from .synapses import Alpha, LinearFilter, Lowpass, Triangle
-from .transforms import Dense, Convolution
+from .transforms import Convolution, Dense, Sparse
 from .utils.logging import log
 from . import dists, exceptions, networks, presets, processes, spa, utils