+++ nengo/builder/signal.py
@@ -3,7 +3,12 @@ from io import StringIO
 import numpy as np
 
 from nengo.exceptions import SignalError
+from nengo.transforms import SparseMatrix
 import nengo.utils.numpy as npext
+
+
+def is_sparse(obj):
+    return isinstance(obj, SparseMatrix) or npext.is_spmatrix(obj)
 
 
 class Signal:
@@ -51,12 +56,26 @@ class Signal:
             assert name
         self._name = name
 
-        initial_value = np.asarray(initial_value)
-        if initial_value.ndim > 0 and base is None:
-            self._initial_value = np.ascontiguousarray(initial_value).view()
+        self._initial_value = initial_value
+        if self.sparse:
+            assert initial_value.ndim == 2
+            assert offset == 0
+            assert base is None
+            if npext.is_spmatrix(initial_value):
+                self._initial_value.data.setflags(write=False)
+
         else:
-            self._initial_value = initial_value.view()
-        self._initial_value.setflags(write=False)
+            # To ensure we do not modify data passed into the signal,
+            # we make a view of the data and mark it as not writeable.
+            # Consumers (like SignalDict) are responsible for making copies
+            # that can be modified, or using the readonly view appropriately.
+            readonly_view = np.asarray(self._initial_value)
+            if readonly_view.ndim > 0 and base is None:
+                readonly_view = np.ascontiguousarray(readonly_view)
+            # Ensure we have a view and aren't modifying the original's flags
+            readonly_view = readonly_view.view()
+            readonly_view.setflags(write=False)
+            self._initial_value = readonly_view
 
         if base is not None:
             assert isinstance(base, Signal) and not base.is_view
@@ -69,22 +88,35 @@ class Signal:
 
     def __getstate__(self):
         state = dict(self.__dict__)
-        v = self._initial_value
-        state['_initial_value'] = (
-            v.shape, v.base, npext.array_offset(v), v.strides)
+
+        if not self.sparse:
+            # For normal arrays, the initial value could be a view on another
+            # signal's data. To make sure we do not make a copy of the data,
+            # we store the underlying metadata in a tuple, which pickle can
+            # inspect to see that v.base is the same in different signals
+            # and avoid serializing it multiple times.
+            v = self._initial_value
+            state['_initial_value'] = (
+                v.shape, v.base, npext.array_offset(v), v.strides)
+
         return state
 
     def __setstate__(self, state):
         for k, v in state.items():
             setattr(self, k, v)
 
-        shape, base, offset, strides = self._initial_value
-        self._initial_value = np.ndarray(
-            shape, buffer=base, offset=offset, strides=strides)
-        self._initial_value.setflags(write=False)
+        if not self.sparse:
+            shape, base, offset, strides = self._initial_value
+            self._initial_value = np.ndarray(
+                shape, buffer=base, offset=offset, strides=strides)
+            self._initial_value.setflags(write=False)
 
     def __getitem__(self, item):
         """Index or slice into array"""
+
+        if self.sparse:
+            raise SignalError("Attempting to create a view of a sparse Signal")
+
         if item is Ellipsis or (
                 isinstance(item, slice) and item == slice(None)):
             return self
@@ -106,11 +138,11 @@ class Signal:
                       base=self.base, offset=offset)
 
     def __repr__(self):
-        return "Signal(%s, shape=%s)" % (self._name, self.shape)
+        return "Signal(name=%s, shape=%s)" % (self._name, self.shape)
 
     @property
     def base(self):
-        """(Signal or None) The base signal, if this signal is a view.
+        """(Signal) The base signal, if this signal is a view.
 
         Linking the two signals with the ``base`` argument is necessary
         to ensure that their live data is also linked.
@@ -130,7 +162,8 @@ class Signal:
     @property
     def elemstrides(self):
         """(int) Strides of data in elements."""
-        return tuple(s // self.itemsize for s in self.strides)
+        return (None if self.sparse
+                else tuple(s // self.itemsize for s in self.strides))
 
     @property
     def initial_value(self):
@@ -153,7 +186,7 @@ class Signal:
     @property
     def itemsize(self):
         """(int) Size of an array element in bytes."""
-        return self.initial_value.itemsize
+        return self.dtype.itemsize
 
     @property
     def name(self):
@@ -167,8 +200,7 @@ class Signal:
     @property
     def nbytes(self):
         """(int) Number of bytes consumed by the signal."""
-        # Equivalent to self.itemsize * self.size
-        return self.initial_value.nbytes
+        return self.itemsize * self.size
 
     @property
     def ndim(self):
@@ -210,7 +242,12 @@ class Signal:
     @property
     def strides(self):
         """(tuple) Strides of data in bytes."""
-        return self.initial_value.strides
+        return None if self.sparse else self.initial_value.strides
+
+    @property
+    def sparse(self):
+        """(bool) Whether the signal is sparse."""
+        return is_sparse(self.initial_value)
 
     def may_share_memory(self, other):
         """Determine if two signals might overlap in memory.
@@ -223,7 +260,8 @@ class Signal:
         other : Signal
             The other signal we are investigating.
         """
-        return np.may_share_memory(self.initial_value, other.initial_value)
+        return (self.is_view or other.is_view) and np.may_share_memory(
+            self.initial_value, other.initial_value)
 
     def reshape(self, *shape):
         """Return a view on this signal with a different shape.
@@ -234,6 +272,10 @@ class Signal:
         Any number of integers can be passed to this method,
         describing the desired shape of the returned signal.
         """
+
+        if self.sparse:
+            raise SignalError("Attempting to create a view of a sparse Signal")
+
         if len(shape) == 1:
             shape = shape[0]  # in case a tuple is passed in
         initial_value = self.initial_value.view()
@@ -291,25 +333,29 @@ class SignalDict(dict):
         return sio.getvalue()
 
     def init(self, signal):
-        """Set up a permanent mapping from signal -> ndarray."""
+        """Set up a permanent mapping from signal -> data."""
         if signal in self:
             raise SignalError("Cannot add signal twice")
 
-        x = signal.initial_value
+        assert isinstance(signal, Signal)
+        data = signal.initial_value
+        if isinstance(data, SparseMatrix):
+            data = data.allocate()
+
         if signal.is_view:
             if signal.base not in self:
                 self.init(signal.base)
 
             # get a view onto the base data
             view = np.ndarray(
-                shape=x.shape, strides=x.strides, offset=signal.offset,
-                dtype=x.dtype, buffer=self[signal.base].data)
-            assert np.array_equal(view, x)
+                shape=data.shape, strides=data.strides, offset=signal.offset,
+                dtype=data.dtype, buffer=self[signal.base].data)
+            assert np.array_equal(view, data)
             view.setflags(write=not signal.readonly)
             dict.__setitem__(self, signal, view)
         else:
-            x = x.view() if signal.readonly else x.copy()
-            dict.__setitem__(self, signal, x)
+            data = data if signal.readonly else data.copy()
+            dict.__setitem__(self, signal, data)
 
     def reset(self, signal):
         """Reset ndarray to the base value of the signal that maps to it"""